+++ apps/web/src/modules/organization/routes.tsx
@@ -14,6 +14,8 @@ import { UnitFormPage } from './shared/UnitFormPage';
 import { branchConfig, departmentConfig, sectionConfig } from './shared/unit-config';
 import { BranchesListPage } from './branches/pages/BranchesListPage';
 import { BranchDetailPage } from './branches/pages/BranchDetailPage';
+import { DepartmentsListPage } from './departments/pages/DepartmentsListPage';
+import { DepartmentDetailPage } from './departments/pages/DepartmentDetailPage';
 import { JobTitlesListPage } from './job-titles/pages/JobTitlesListPage';
 import { JobTitleDetailPage } from './job-titles/pages/JobTitleDetailPage';
 import { JobTitleFormPage } from './job-titles/pages/JobTitleFormPage';
@@ -69,7 +71,7 @@ export default function OrganizationRoutes(): JSX.Element {
             </RequirePermission>
           }
         >
-          <Route index element={<UnitListPage config={departmentConfig} />} />
+          <Route index element={<DepartmentsListPage />} />
           <Route
             path="new"
             element={
@@ -78,7 +80,7 @@ export default function OrganizationRoutes(): JSX.Element {
               </RequirePermission>
             }
           />
-          <Route path=":id" element={<UnitDetailPage config={departmentConfig} />} />
+          <Route path=":id" element={<DepartmentDetailPage />} />
           <Route
             path=":id/edit"
             element={